alot of work on processes and threads

+++ src/arch/gdt.hpp
@@ -1,10 +1,12 @@
 #pragma once
 
+#include <stdint.h>
+
 namespace arch::gdt {
 
-    constexpr int kernel_code = 8;
-    constexpr int kernel_data = 16;
-    constexpr int user_data = (24 | 3);
-    constexpr int user_code = (32 | 3);
+    constexpr uint16_t kernel_code = 8u;
+    constexpr uint16_t kernel_data = 16u;
+    constexpr uint16_t user_data = (24u | 3u);
+    constexpr uint16_t user_code = (32u | 3u);
 
 }
